Adding patch impl.

+++ AeroGearSync/BackupShadowDocument.swift
@@ -1,10 +1,10 @@
 import Foundation
 
 public class BackupShadowDocument<T> {
-    public let version: UInt64
+    public let version: Int
     public let shadowDocument: ShadowDocument<T>
     
-    public init(version: UInt64, shadowDocument: ShadowDocument<T>) {
+    public init(version: Int, shadowDocument: ShadowDocument<T>) {
         self.version = version
         self.shadowDocument = shadowDocument
     }
