+++ AeroGearSync/ShadowDocument.swift
@@ -1,11 +1,11 @@
 import Foundation
 
 public class ShadowDocument<T> {
-    public let clientVersion: UInt64
-    public let serverVersion: UInt64
+    public let clientVersion: Int
+    public let serverVersion: Int
     public let clientDocument: ClientDocument<T>
 
-    public init(clientVersion: UInt64, serverVersion: UInt64, clientDocument: ClientDocument<T>) {
+    public init(clientVersion: Int, serverVersion: Int, clientDocument: ClientDocument<T>) {
         self.clientVersion = clientVersion
         self.serverVersion = serverVersion
         self.clientDocument = clientDocument
